# postman/specs/openapi.yaml
openapi: 3.0.3
info:
  title: FameTech Developer API
  version: "2.0.0"
  description: |
    The public developer API for FameTech — Ghana digital services
    (MTN/Telecel/AirtelTigo data bundles, airtime, results-checker vouchers,
    MTN AFA registration, utility bill payments, and SMS).

    **v2 is the only live API** — v1 was fully ported and then retired
    (owner decision, 2026-08-31; all developers were notified and confirmed
    before the cutover). Every endpoint below lives under `/api/v2`.

    ### Response envelope
    Every endpoint returns one of two shapes:
    - Success: `{ "success": true, "data": { ... }, "meta": { "timestamp": "...", "version": "v2" } }`
    - Error: `{ "success": false, "error": { "code": 400, "message": "..." } }`

    ### Ghana only
    This platform serves Ghanaian mobile networks and services exclusively
    (MTN, Telecel, AirtelTigo). There is no coverage outside Ghana.

    ### Network number validation
    We validate a recipient/beneficiary number's network from its prefix
    before submitting an order. As of this writing:
    - **MTN**: 024, 025, 053, 054, 055, 059
    - **Telecel**: 020, 050
    - **AirtelTigo**: 026, 027, 056, 057

    These prefix-to-network mappings are **not guaranteed to stay fixed** —
    Ghanaian operators occasionally get reassigned or new ranges opened by
    the regulator. Don't hardcode this list into your own client-side
    validation as a permanent source of truth; treat our API's response as
...

...
servers:
...
security:
  - ApiKeyAuth: []
...
paths:
...
  /sms/send:
    post:
      tags: [SMS]
      operationId: sendSms
      summary: Send an SMS to one or many recipients
      description: |
        Requires an SMS-type API key (`requireKeyType(auth, 'sms')`) — a
        standard or commission key is rejected with 403. Small sends
        (≤500 recipients) dispatch immediately; larger sends are queued and
        processed within a minute.
      security:
        - SmsApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [message, recipients]
              properties:
                message: { type: string, minLength: 3, maxLength: 1000 }
                recipients:
                  oneOf:
                    - type: string
                    - type: array
                      items: { type: string }
                      maxItems: 10000
                  description: "Ghana number(s), 0XXXXXXXXX or 233XXXXXXXXX. Duplicates removed automatically."
                sender: { type: string, description: "Must be one of your approved sender IDs or a pool sender — see GET /sms/senders" }
                reference: { type: string, maxLength: 100, description: "Optional idempotency key" }
      responses:
        "200":
          description: Sent or queued
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          campaignId: { type: string, format: uuid }
                          status: { type: string, example: completed }
                          recipients: { type: integer }
                          segments: { type: integer }
                          creditsCharged: { type: integer }
                          sender: { type: string }
                          sent: { type: integer }
                          failed: { type: integer }
                          balance: { type: integer }
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "402":
          description: Insufficient SMS credits
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiErrorEnvelope' }
        "403": { $ref: '#/components/responses/Forbidden' }
        "429": { $ref: '#/components/responses/RateLimited' }
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: |
        Your standard API key, sent RAW (not `Bearer <key>`) — e.g.
        `Authorization: kf_live_...`. Standard keys work on every endpoint
        except the SMS, Airtime, and Utility Bills (Commission Services)
        groups, which require their own key type (see SmsApiKeyAuth /
        CommissionApiKeyAuth).
    SmsApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: An SMS-type key (`kf_sms_live_...`). A standard or commission key is rejected here with 403.
    CommissionApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: A Commission Services key (`kf_cs_live_...`), required for utility bill AND airtime endpoints. No shop required — commission is paid into a dedicated commission wallet, separate from shop earnings. A standard key is rejected here with 403, and this key is rejected with 403 everywhere else.
...
  responses:
    BadRequest:
      description: Invalid request body or parameters
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiErrorEnvelope' }
    Unauthorized:
      description: Missing or invalid API key
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiErrorEnvelope' }
    Forbidden:
      description: Wrong key type for this endpoint, key pending/revoked, role not allowed, or account suspended
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiErrorEnvelope' }
...
    RateLimited:
      description: Rate limit exceeded for this endpoint
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiErrorEnvelope' }
  schemas:
    ApiSuccessEnvelope:
      type: object
      required: [success, data, meta]
      properties:
        success: { type: boolean, enum: [true] }
        data: { type: object }
        meta:
          type: object
          properties:
            timestamp: { type: string, format: date-time }
            version: { type: string, enum: [v2] }
    ApiErrorEnvelope:
      type: object
      required: [success, error]
      properties:
        success: { type: boolean, enum: [false] }
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: integer }
            message: { type: string }
...